+++ R/er_meas_scores.R
@@ -33,7 +33,7 @@ er_meas_score <- function(B_SOILTYPE_AGR, B_AER_CBS,B_AREA,
   eco_app = b_lu_arable_er = b_lu_productive_er = b_lu_cultivated_er = NULL
   er_total = er_climate = er_soil = er_measure = er_water = er_landscape = er_biodiversity = NULL
   reward_cf = regio_factor = . = er_cf = statcode = NULL
-  code = choices = NULL
+  code = choices = area_fr = NULL
   
   # Load bbwp_parms
   bbwp_parms <- BBWPC::bbwp_parms
@@ -61,6 +61,7 @@ er_meas_score <- function(B_SOILTYPE_AGR, B_AER_CBS,B_AREA,
   
   # filter out measures already receiving points from crop rotation 
   dt.meas.taken <- dt.meas.taken[!(grepl('EB1$|EB2$|EB3$|EB8|EB9',eco_id) & level == 'field'),]
+  dt.meas.taken <- dt.meas.taken[level != 'farm']
   
   # add bbwp table for financial reward correction factor per AER
   dt.er.reward <- as.data.table(BBWPC::er_aer_reward)
@@ -100,10 +101,11 @@ er_meas_score <- function(B_SOILTYPE_AGR, B_AER_CBS,B_AREA,
               dt.meas.eco, 
               by = c('B_LU_BRP','eco_id'),
               all.x = TRUE)
-  dt[is.na(eco_app),eco_app := 0]
+  dt[is.na(eco_app) & !grepl('EG20',eco_id),eco_app := 0]
+  dt[is.na(eco_app) & grepl('EG20',eco_id), eco_app := 1]
   
   # measures that apply to crops cultivated after main crop (vanggewassen en groenbemesters) and FAB-stroken are applicable on all main crops 
-  dt <- dt[eco_id  == "EB17|EB10|EB23" & B_LU_BBWP == "gras_tijdelijk|rustgewas|rooivrucht|groenten|bollensierteelt|boomfruitteelt|mais|eiwitgewas", eco_app := 1]
+  dt[grepl("EB17|EB10|EB23", eco_id) & grepl("gras_tijdelijk|rustgewas|rooivrucht|groenten|bollensierteelt|boomfruitteelt|mais|eiwitgewas",B_LU_BBWP), eco_app := 1]
   
   # set scores to zero when measures are not applicable given the crop type
   
@@ -131,12 +133,13 @@ er_meas_score <- function(B_SOILTYPE_AGR, B_AER_CBS,B_AREA,
     dt[eco_app == 0, c(cols) := 0]
     
     # set the score to zero when not applicable as arable/productive/cultivated measure
-    dt[B_LU_ARABLE_ER  == TRUE & b_lu_arable_er  == 0, c(cols) := 0]
-    dt[B_LU_PRODUCTIVE_ER == TRUE & b_lu_productive_er == 0, c(cols) := 0]
-    dt[B_LU_CULTIVATED_ER  == TRUE & b_lu_cultivated_er == 0, c(cols) := 0]
-    dt[B_LU_ARABLE_ER  == FALSE & b_lu_arable_er  == 1, c(cols) := 0]
-    dt[B_LU_PRODUCTIVE_ER == FALSE & b_lu_productive_er == 1, c(cols) := 0]
-    dt[B_LU_CULTIVATED_ER  == FALSE & b_lu_cultivated_er == 1, c(cols) := 0]
+    # ensure that measure non-productive area is always applicable
+    dt[B_LU_ARABLE_ER  == TRUE & b_lu_arable_er  == 0 & !grepl('EG20',eco_id), c(cols) := 0]
+    dt[B_LU_PRODUCTIVE_ER == TRUE & b_lu_productive_er == 0 & !grepl('EG20',eco_id), c(cols) := 0]
+    dt[B_LU_CULTIVATED_ER  == TRUE & b_lu_cultivated_er == 0 & !grepl('EG20',eco_id), c(cols) := 0]
+    dt[B_LU_ARABLE_ER  == FALSE & b_lu_arable_er  == 1 & !grepl('EG20',eco_id), c(cols) := 0]
+    dt[B_LU_PRODUCTIVE_ER == FALSE & b_lu_productive_er == 1 & !grepl('EG20',eco_id), c(cols) := 0]
+    dt[B_LU_CULTIVATED_ER  == FALSE & b_lu_cultivated_er == 1 & !grepl('EG20',eco_id), c(cols) := 0]
     
   # set the score and profit to zero when the measure is not applicable given sector or soil type
   
@@ -202,7 +205,7 @@ er_meas_score <- function(B_SOILTYPE_AGR, B_AER_CBS,B_AREA,
     dt[grepl('^EB12',eco_id) & B_AREA_REL > 71 & B_AREA_REL <= 85 & er_total > 0, c(cols.sel) := Map('+',mget(cols.sel),cols.ad1)]
     dt[grepl('^EB12',eco_id) & B_AREA_REL > 85 & er_total > 0, c(cols.sel) := Map('+',mget(cols.sel),cols.ad2)]
     
-    # measure EB13. no till and reduced till
+    # measure EB13. no till and reduced till 
     cols.ad1 <- c(1,1,0,0,0,0)
     cols.ad2 <- c(2,2,0,0,0,0)
     dt[grepl('^EB13',eco_id), B_AREA_REL := sum(B_AREA) * 100 / dt.farm$area_arable]
@@ -218,12 +221,17 @@ er_meas_score <- function(B_SOILTYPE_AGR, B_AER_CBS,B_AREA,
     dt[grepl('^EG11',eco_id) & B_AREA_REL > 25 & B_AREA_REL <= 50 & er_total > 0, c(cols.sel) := Map('+',mget(cols.sel),cols.ad1)]
     dt[grepl('^EG11',eco_id) & B_AREA_REL > 50 & er_total > 0, c(cols.sel) := Map('+',mget(cols.sel),cols.ad2)]
     
-    # measure EG20. Not productive land
+    # measure EG20. Not productive land # Does not work in BBWP now # farmers indicate percentage themselves
+    # reken uit per perceel
     cols.ad1 <- c(1,0,1,3,5,0)
     cols.ad2 <- c(2,0,2,6,10,0)
-    dt[grepl('^EG20',eco_id),B_AREA_REL := sum(B_AREA) * 100 / dt.farm$area_farm]
-    dt[grepl('^EG20',eco_id) & B_AREA_REL <= 5 & er_total > 0, c(cols.sel) := 0]
-    dt[grepl('^EG20',eco_id) & B_AREA_REL > 7 & B_AREA_REL <= 9 & er_total > 0, c(cols.sel) := Map('+',mget(cols.sel),cols.ad1)]
+    dt[,area_fr := 0]
+    dt[grepl('^EG20A',eco_id), area_fr := 0.04 + 0.03]
+    dt[grepl('^EG20B',eco_id), area_fr := 0.04 + 0.05]
+    dt[grepl('^EG20C',eco_id), area_fr := 0.04 + 1.00]
+    dt[grepl('^EG20',eco_id),B_AREA_REL := sum(B_AREA * area_fr) * 100 / dt.farm$area_farm]
+    dt[grepl('^EG20',eco_id) & B_AREA_REL < 5 & er_total > 0, c(cols.sel) := 0]
+    dt[grepl('^EG20',eco_id) & B_AREA_REL > 7 & B_AREA_REL < 9 & er_total > 0, c(cols.sel) := Map('+',mget(cols.sel),cols.ad1)] 
     dt[grepl('^EG20',eco_id) & B_AREA_REL > 9 & er_total > 0, c(cols.sel) := Map('+',mget(cols.sel),cols.ad2)]
     
     # measure EG13. inzet baggerspuit (check na update maatregelentabel, EG13 kan 1 keer per perceel voorkomen)
@@ -271,7 +279,7 @@ er_meas_score <- function(B_SOILTYPE_AGR, B_AER_CBS,B_AREA,
     dt2[, oid := frank(-total, ties.method = 'first',na.last = 'keep'),by = c('id','bbwp_conflict')]
     dt2[oid > 1, c(cols) := 0]
     
-  # calculate the weighed average ER score (points/ ha) for the whole farm due to measures taken
+  # calculate the total ER score (points/ ha) per field due to measures taken
   dt.field <- dt2[,lapply(.SD,sum), .SDcols = cols, by = 'id']
     
   # select total reward per field which is equal to the reward from the measure with the highest er_euro_ha (euro / ha)