+++ dev/test_astrid2.R
@@ -1,3 +1,8 @@
+library(data.table)
+library(BBWPC)
+
+# LOCATION TO STORE CSV TESTS
+loc <- 'C:/R_packages_Tessa'
 
 # get internal table with measures
 dt.measures <- as.data.table(BBWPC::bbwp_measures)
@@ -9,7 +14,7 @@ B_GWL_CLASS = c('GtIII')
 A_P_SG = c(12)
 B_SLOPE_DEGREE = c(1.5)
 B_AER_CBS = c('LG01')
-B_LU_BBWP = c('eiwitgewas')
+B_LU_BBWP = c('gras_tijdelijk') #aangepast, was "eiwitgewas". Eiwitgewas voegde overal de punten van EB2 toe.
 B_LU_BRP = c(800)
 B_LU_ARABLE_ER = c(T)
 B_LU_PRODUCTIVE_ER = c(T)
@@ -19,10 +24,13 @@ D_SA_W = c(0.5)
 B_AREA = c(10)
 farmscore = 100
 medalscore = "gold"
-measures = measures
+measures = dt.measures
 sector = c('arable')
 output = 'scores'
-list1 = list2 = list3 = list()
+
+# list 1 bevat field and farm score, averaged on farm level
+# list 2 bevat only farm measures (all measures for crop rotation plus level==farm)
+list1 = list2 = list()
 
 for(i in 1:nrow(dt.measures)){
   
@@ -40,104 +48,23 @@ for(i in 1:nrow(dt.measures)){
                            B_LU_ARABLE_ER = B_LU_ARABLE_ER,
                            B_LU_PRODUCTIVE_ER = B_LU_PRODUCTIVE_ER,
                            B_LU_CULTIVATED_ER = B_LU_CULTIVATED_ER,
-                           B_CT_SOIL =aim$B_CT_SOIL, 
-                           B_CT_WATER = aim$B_CT_WATER,
-                           B_CT_CLIMATE = aim$B_CT_CLIMATE,
-                           B_CT_BIO = aim$B_CT_BIO,
-                           B_CT_LANDSCAPE = aim$B_CT_LANDSCAPE, 
                            measures = measures, sector)
   list1[[i]] <- copy(cbind(field,aim))
   
-  # get original scores for crop rotation and field level of measures
-  
-  # check and update the measure table
-  dt.er.meas <- bbwp_check_meas(measures, eco = TRUE, score = TRUE)
-  
-  # get internal table with importance of environmental challenges
-  dt.er.scoring <- as.data.table(BBWPC::er_scoring)
-  setnames(dt.er.scoring,gsub('cf_','',colnames(dt.er.scoring)))
-  dt.er.urgency <- melt(dt.er.scoring[type=='urgency'],
-                        id.vars='soiltype',
-                        measure.vars = c('soil', 'water', 'climate',  'biodiversity', 'landscape'),
-                        variable.name = 'indicator',
-                        value.name = 'urgency')
-  
-  # collect data in one data.table
-  dt <- data.table(id = 1,
-                   B_SOILTYPE_AGR = B_SOILTYPE_AGR,
-                   B_LU_BBWP = B_LU_BBWP,
-                   B_LU_BRP = B_LU_BRP,
-                   B_LU_ARABLE_ER = B_LU_ARABLE_ER,
-                   B_LU_PRODUCTIVE_ER = B_LU_PRODUCTIVE_ER,
-                   B_LU_CULTIVATED_ER = B_LU_CULTIVATED_ER,
-                   B_AER_CBS = B_AER_CBS,
-                   B_AREA = B_AREA,
-                   B_CT_SOIL = aim$B_CT_SOIL, 
-                   B_CT_WATER = aim$B_CT_WATER,
-                   B_CT_CLIMATE = aim$B_CT_CLIMATE,
-                   B_CT_BIO = aim$B_CT_BIO,
-                   B_CT_LANDSCAPE = aim$B_CT_LANDSCAPE
-  )
-  
-  # columns with the Ecoregelingen ranks
-  cols <- c('soil','water','biodiversity','climate','landscape')
-  
   # add the generic farm score as baseline
   # this gives the averaged ER score based on the crops in crop rotation plan
-  dt.farm <- er_croprotation(B_SOILTYPE_AGR = dt$B_SOILTYPE_AGR,
-                             B_LU_BBWP = dt$B_LU_BBWP,
+  dt.farm <- er_croprotation(B_SOILTYPE_AGR = B_SOILTYPE_AGR,
+                             B_LU_BBWP = B_LU_BBWP,
                              B_LU_BRP = B_LU_BRP,
                              B_LU_ARABLE_ER = B_LU_ARABLE_ER,
                              B_LU_PRODUCTIVE_ER = B_LU_PRODUCTIVE_ER,
                              B_LU_CULTIVATED_ER = B_LU_CULTIVATED_ER,
-                             B_AER_CBS = dt$B_AER_CBS,
-                             B_AREA = dt$B_AREA,
-                             B_CT_SOIL = dt$B_CT_SOIL,
-                             B_CT_WATER = dt$B_CT_WATER,
-                             B_CT_CLIMATE = dt$B_CT_CLIMATE,
-                             B_CT_BIO = dt$B_CT_BIO,
-                             B_CT_LANDSCAPE = dt$B_CT_LANDSCAPE,
+                             B_AER_CBS = B_AER_CBS,
+                             B_AREA = B_AREA,
                              measures = measures,
                              sector = sector)
   
   list2[[i]] <- copy(dt.farm)
-  
-  # calculate the change in opportunity indexes given the measures taken
-  
-  # column names for impact of measures on the five indexes (do not change order)
-  # these are not yet converted to a 0-1 scale
-  
-  # set colnames for the impact of measures
-  mcols <- c('D_MEAS_BIO', 'D_MEAS_CLIM', 'D_MEAS_LAND', 'D_MEAS_SOIL', 'D_MEAS_WAT','S_ER_REWARD')
-  
-  # calculate the total score per indicator 
-  if(nrow(dt.er.meas) > 0){
-    
-    # calculate
-    dt.meas.impact <- er_meas_score(B_SOILTYPE_AGR = B_SOILTYPE_AGR, 
-                                    B_LU_BBWP = B_LU_BBWP,
-                                    B_LU_BRP = B_LU_BRP,
-                                    B_LU_ARABLE_ER = B_LU_ARABLE_ER,
-                                    B_LU_PRODUCTIVE_ER = B_LU_PRODUCTIVE_ER,
-                                    B_LU_CULTIVATED_ER = B_LU_CULTIVATED_ER,
-                                    B_AER_CBS = B_AER_CBS,
-                                    B_AREA = B_AREA,
-                                    measures = measures, 
-                                    sector = sector)
-    
-    # merge with dt
-    dt <- merge(dt,dt.meas.impact,by='id')
-    
-    # set NA to zero
-    dt[,c(mcols) := lapply(.SD,function(x) fifelse(is.na(x),0,x)),.SDcols = mcols]
-    
-  } else {
-    
-    # set impact of management to zero when no measures are applied
-    dt[,c(mcols) := list(0,0,0,0,0,0)]
-  }
-  
-  list3[[i]] <- copy(dt)
   
   # print
   print(i)
@@ -145,14 +72,12 @@ for(i in 1:nrow(dt.measures)){
 
 # field scores and aim
 dt1 <- rbindlist(list1)
-# dt.farm score used within field score
+# dt.farm score for all crop rotation measures and measures where level == farm
 dt2 <- rbindlist(list2)
-# field scores
-dt3 <- rbindlist(list3)
-#fwrite(dt2,"dev/farm_800_klei_LG01_akkerbouw.csv")
-#fwrite(dt3,"dev/field_800_klei_LG01_akkerbouw.csv")
 
-
+# write as csv
+#(dt2,paste0(loc, "farm_800_klei_LG01_akkerbouw.csv"))
+#fwrite(dt1,paste0(loc, "field_800_klei_LG01_akkerbouw.csv"))
 
 ### total farm with three fields ###
 
@@ -179,7 +104,7 @@ dt3 <- rbindlist(list3)
   sector = c('dairy')
   output = 'scores'
   
-  measures <- rbind(data.table(id = 1, dt.measures[grepl('B138',bbwp_id)]),
+  measures <- rbind(data.table(id = 1, dt.measures[grepl('B138',bbwp_id)]),#EG16B
                     data.table(id = 2, dt.measures[grepl('EG2B$',eco_id)]),
                     data.table(id = 3, dt.measures[grepl('EG14|EG18',eco_id)]))
   
@@ -204,8 +129,12 @@ dt3 <- rbindlist(list3)
                                 medalscore = medalscore,
                                 output = 'scores'
                                 )
+  #quick fix under bedrijfsscore1 farm must also get variable field_id, otherwise dimensions don't match
+  bedrijfsscore1$farm$field_id <- bedrijfsscore1$fields$field_id
   
-  write.csv2(bedrijfsscore1,"dev/bedrijf1.csv")
+#  fwrite(bedrijfsscore1,paste0(loc, "bedrijf1.csv"))
+  
+#  fwrite(bedrijfsscore1$fields,paste0(loc, "bedrijf1_fields.csv"))
   
 ### total farm with four fields ###
   
@@ -260,5 +189,64 @@ dt3 <- rbindlist(list3)
                                output = 'scores'
   )
   
-  write.csv2(bedrijfsscore2,"dev/bedrijf2.csv")
+  #quick fix under bedrijfsscore1 farm must also get variable field_id, otherwise dimensions don't match
+  bedrijfsscore2$farm$field_id <- bedrijfsscore2$fields$field_id
   
+#  fwrite(bedrijfsscore2,paste0(loc, "bedrijf2.csv"))  
+#  fwrite(bedrijfsscore2$fields,paste0(loc, "bedrijf2_field.csv"))
+  
+  
+  ## Bedrijf 3aanpassingen: Boerennatuur willen combinatie van binnen een zone 
+  #2x dekzand, maar andere maatregelen per perceel
+  #met niet=prodcutief land
+
+  aantal = 2
+  # input data
+  B_SOILTYPE_AGR = c('dekzand','dekzand')
+  B_GWL_CLASS = c('GtIII','GtIII')
+  A_P_SG = c(12,12)
+  B_SLOPE_DEGREE = c(1.5,1.5)
+  B_AER_CBS = c('LG01','LG01')
+  B_LU_BBWP = c('eiwitgewas','eiwitgewas')
+  B_LU_BRP = c(258,258)
+  B_LU_ARABLE_ER = c(T,F)
+  B_LU_PRODUCTIVE_ER = c(F,T)
+  B_LU_CULTIVATED_ER = c(F,T)
+  M_DRAIN = c(T,T)
+  D_SA_W = c(0.5,0.5)
+  B_AREA = c(10,10)
+  farmscore = 100
+  medalscore = "gold"
+  measures = measures
+  sector = c('arable')
+  output = 'scores'
+  
+  measures <- rbind(data.table(id = 1, dt.measures[grepl('EG16B|EB13A',bbwp_id)]),
+                    data.table(id = 2, dt.measures[grepl('EG20',eco_id)])
+                    )
+  
+  measures$bbwp_status <- 'hello check'
+  
+  bedrijfsscore3 <- ecoregeling(B_SOILTYPE_AGR = B_SOILTYPE_AGR,
+                                B_GWL_CLASS = B_GWL_CLASS,
+                                B_AER_CBS = B_AER_CBS,
+                                A_P_SG = A_P_SG,
+                                B_SLOPE_DEGREE = B_SLOPE_DEGREE,
+                                B_AREA = B_AREA,
+                                B_LU_BBWP = B_LU_BBWP,
+                                B_LU_BRP = B_LU_BRP,
+                                B_LU_ARABLE_ER = B_LU_ARABLE_ER,
+                                B_LU_PRODUCTIVE_ER = B_LU_PRODUCTIVE_ER,
+                                B_LU_CULTIVATED_ER = B_LU_CULTIVATED_ER,
+                                M_DRAIN = M_DRAIN, 
+                                D_SA_W = D_SA_W,
+                                measures = measures, 
+                                sector = sector,
+                                farmscore = farmscore,
+                                medalscore = medalscore,
+                                output = 'scores'
+  )
+  #quick fix under bedrijfsscore1 farm must also get variable field_id, otherwise dimensions don't match
+  bedrijfsscore3$farm$field_id <- bedrijfsscore3$fields$field_id
+  
+#  fwrite(bedrijfsscore3,paste0(loc, "bedrijf3.csv"))