+++ tests/testthat/test-er_croprotation.R
@@ -2,22 +2,16 @@
 require(testthat)
 
   # default input for testing
-  B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei','veen')
-  B_LU_BBWP = rep('gras_permanent',4)
-  B_LU_BRP = rep(265,4)
-  B_LU_ARABLE_ER = c(T,T,T,T)
-  B_LU_PRODUCTIVE_ER = c(T,T,T,T)
-  B_LU_CULTIVATED_ER = c(T,T,T,T)
-  B_AER_CBS = c('Bouwhoek en Hogeland','LG14','LG12','Westelijk Holland')
-  B_AREA = c(45,18,0.8,6)
-  erf <- er_farm_aim(B_SOILTYPE_AGR,B_AREA)
-  B_CT_SOIL = erf$B_CT_SOIL
-  B_CT_WATER = erf$B_CT_WATER
-  B_CT_CLIMATE = erf$B_CT_CLIMATE
-  B_CT_BIO = erf$B_CT_BIO
-  B_CT_LANDSCAPE = erf$B_CT_LANDSCAPE
-  measures = NULL
-  sector = 'dairy'
+  # B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei','veen')
+  # B_LU_BBWP = rep('gras_permanent',4)
+  # B_LU_BRP = rep(265,4)
+  # B_LU_ARABLE_ER = c(T,T,T,T)
+  # B_LU_PRODUCTIVE_ER = c(T,T,T,T)
+  # B_LU_CULTIVATED_ER = c(T,T,T,T)
+  # B_AER_CBS = c('Bouwhoek en Hogeland','LG14','LG12','Westelijk Holland')
+  # B_AREA = c(45,18,0.8,6)
+  # measures = NULL
+  # sector = 'dairy'
 
 # run example 1 without any measures taken
 test <- er_croprotation(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei','veen'),
@@ -28,11 +22,6 @@ test <- er_croprotation(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei','vee
                         B_LU_CULTIVATED_ER = c(T,T,T,T),
                         B_AER_CBS = c('Bouwhoek en Hogeland','LG14','LG12','Westelijk Holland'),
                         B_AREA = c(45,18,0.8,6),
-                        B_CT_SOIL = B_CT_SOIL, 
-                        B_CT_WATER = B_CT_WATER,
-                        B_CT_CLIMATE = B_CT_CLIMATE,
-                        B_CT_BIO = B_CT_BIO,
-                        B_CT_LANDSCAPE = B_CT_LANDSCAPE,
                         measures = NULL,
                         sector = 'dairy'
                         )
@@ -77,11 +66,6 @@ test <- er_croprotation(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei','vee
                         B_LU_CULTIVATED_ER = c(T,T,T,T),
                         B_AER_CBS = c('Bouwhoek en Hogeland','LG14','LG12','Westelijk Holland'),
                         B_AREA = c(45,18,0.8,6),
-                        B_CT_SOIL = B_CT_SOIL, 
-                        B_CT_WATER = B_CT_WATER,
-                        B_CT_CLIMATE = B_CT_CLIMATE,
-                        B_CT_BIO = B_CT_BIO,
-                        B_CT_LANDSCAPE = B_CT_LANDSCAPE,
                         measures = measures,
                         sector = 'dairy'
                      )
@@ -103,7 +87,7 @@ test_that("check er_croprotation", {
 test_that("check er_croprotation", {
   expect_equal(
     object = as.numeric(test),
-    expected = c(1,67.35,9.39,58.55,33.53,52.81,222,113.4),
+    expected = c(1,77.3,16.7,68.1,28.9,61.1,252.2,1768.2),
     tolerance = 0.1,
     ignore_attr = TRUE)
 })